modify indentation and add errno
